requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0 